refactor: move Footer to /_components
