+++ apps/web/app/articles/_features/index.ts
@@ -1,2 +1,1 @@
 export { ArticleList } from "./ArticleList";
-export { Footer } from "./Footer";
